SubMenu.java is added.
SubMenu class is the menu that appears after a quiz finishes. The menu gives a option of repeating the quiz or going back to the main menu. If the user get 9 out of 10 words or more, the user gets a option of going to the next level or playing the reward video. At the top of the menu, the score of the quiz is displayed.